requirements.txt
docling
ollama
requests
python-dotenv
pathlib
